dockerfiles/sidecar-measurer/requirements.txt
redis
fastapi
uvicorn
python-dotenv
web3==5.31.3
hexbytes
